Update dependencies & convert to .NET Standard 1.3
 * Make this project support .NET Standard 1.3.

 * Upgrade to Moq 4.7 and NUnit 3.6 (among other reasons because of
   the above). This however means we can no longer target .NET 4.0,
   but only .NET 4.5.

 * The update to NUnit 3.6 requires changing `[ExpectedException]`
   (which has been obsoleted) to `Assert.Throws`.

 * Separate unit tests into separate project, so that the main
   package does not have to depend on NUnit.

 * Remove the build process dependency on NAnt. Change the build to
   vanilla Visual Studio 2017. This can also build the NuGet package.

 * Fix indentation in `BlogPresenterTests.cs`.

 * Bump version to 2.0.0. (Dropping support for .NET 4.0 is a
   breaking change, so major version needs to go up.)
