--- Moq.Sequences.build
@@ -1,64 +1,0 @@
-<project default="build">
-
-  <property name="configuration"     value="Release"/>
-  <property name="msbuild.verbosity" value="normal"/>
-  <property name="bin.dir"           value="src\bin\${configuration}" dynamic="true" />
-  <property name="log.dir"           value="${path::get-full-path('build\log')}\${configuration}" dynamic="true" />
-  <property name="lib.dir"           value="src\lib"/>
-
-  <property name="nunit.version"     value="2.5.10.11092"/>
-  <property name="nunit.exe"         value="src\packages\NUnit.${nunit.version}\tools\nunit-console.exe"/>
-  <property name="nunit.output"      value="${log.dir}\UnitTestReport.xml" dynamic="true"/>
-  <property name="sequence.dll"      value="Moq.Sequences.dll"/>
-  <property name="nunit.args"        value='${sequence.dll} /nologo /nodots /noshadow /xml="${nunit.output}" /framework:net-4.0' dynamic="true"/>
-
-  <target name="build" depends="clean compile test" description="Cleans, builds and tests Moq.Sequences.dll."/>
-
-  <target name="clean" description="Removes all built files and folders.">
-    <property name="msbuild.target" value="Clean"/>
-    <call target="msbuild"/>
-    <delete dir="build" failonerror="false"/>
-  </target>
-
-  <target name="compile" description="Compiles Moq.Sequences.dll." >
-    <property name="msbuild.target" value="Rebuild"/>
-    <call target="msbuild"/>
-  </target>
-
-  <target name="test" description="Runs the NUnit tests." >
-    <exec program="${nunit.exe}" commandline="${nunit.args}" workingdir="${bin.dir}"/>
-  </target>
-
-  <target name="publish" depends="build" description="Publishes a new package to NuGet." >
-    <delete>
-      <fileset basedir="src" >
-        <include name="*.nupkg"/>
-      </fileset>
-    </delete> 
-    <exec program="nuget" commandline="pack -sym Sequences.csproj -Prop Configuration=Release" workingdir="src"/>
-    <exec program="nuget" commandline="push Moq.Sequences.1.0.1.0.nupkg" workingdir="src"/>
-  </target>
-
-  <target name="debug" description="Causes any following targets to use the debug configuration." >
-    <property name="configuration" value="Debug"/>
-  </target>
-
-  <target name="verbose" description="Turns on additional logging for targets.">
-    <property name="msbuild.verbosity" value="detailed"/>
-    <property name="ncover.verbosity" value="Verbose"/>
-  </target>
-
-  <target name="msbuild" >
-    <mkdir dir="${log.dir}"/>
-    <exec program="${framework::get-tool-path('msbuild.exe')}" verbose="true" workingdir="src" >
-      <arg value="/nologo"/>
-      <arg value="/noconsolelogger"/>
-      <arg value="/fileLogger"/>
-      <arg value="/fileLoggerParameters:LogFile=${log.dir}\${msbuild.target}.log;Verbosity=${msbuild.verbosity}"/>
-      <arg value="/target:${msbuild.target}"/>
-      <arg value="/property:Configuration=${configuration}"/>
-      <arg value="/property:SolutionDir=${project::get-base-directory()}\\"/>
-    </exec>
-  </target>
-
-</project>
