--- src/.gitignore
@@ -1,7 +1,0 @@
-*ReSharper*
-bin/
-obj/
-packages/
-*.csproj.user
-*.suo
-*.nupkg
