+++ src/Moq.Sequences.sln
@@ -1,0 +1,28 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.26403.7
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Moq.Sequences", "Moq.Sequences\Moq.Sequences.csproj", "{09924001-C8F8-450D-84F8-661265DBE601}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Moq.Sequences.Tests", "Moq.Sequences.Tests\Moq.Sequences.Tests.csproj", "{031B8197-7B35-43B4-8077-850AD92407D1}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{09924001-C8F8-450D-84F8-661265DBE601}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{09924001-C8F8-450D-84F8-661265DBE601}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{09924001-C8F8-450D-84F8-661265DBE601}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{09924001-C8F8-450D-84F8-661265DBE601}.Release|Any CPU.Build.0 = Release|Any CPU
+		{031B8197-7B35-43B4-8077-850AD92407D1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{031B8197-7B35-43B4-8077-850AD92407D1}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{031B8197-7B35-43B4-8077-850AD92407D1}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{031B8197-7B35-43B4-8077-850AD92407D1}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
