+++ src/Moq.Sequences/Moq.Sequences.csproj
@@ -1,0 +1,25 @@
+﻿<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <TargetFrameworks>netstandard1.3;net45</TargetFrameworks>
+    <Version>2.0.0</Version>
+    <Authors>Declan Whelan</Authors>
+    <Copyright>© Declan Whelan 2010</Copyright>
+    <Description>Supports checking that Moq setups are executed in sequence.</Description>
+    <Company />
+    <PackageLicenseUrl>https://github.com/dwhelan/Moq-Sequences/blob/master/License.txt</PackageLicenseUrl>
+    <PackageProjectUrl>https://github.com/dwhelan/Moq-Sequences</PackageProjectUrl>
+    <RepositoryUrl>https://github.com/dwhelan/Moq-Sequences.git</RepositoryUrl>
+    <RepositoryType>Git</RepositoryType>
+    <PackageReleaseNotes>Added NuGet support.</PackageReleaseNotes>
+    <PackageTags>moq mock .net45 netstandard1.3</PackageTags>
+    <AssemblyVersion>2.0.0.0</AssemblyVersion>
+    <FileVersion>2.0.0.0</FileVersion>
+    <GeneratePackageOnBuild>True</GeneratePackageOnBuild>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Moq" Version="4.7.0" />
+  </ItemGroup>
+
+</Project>
