+++ src/Moq.Sequences/SequenceException.cs
@@ -1,0 +1,11 @@
+using System;
+
+namespace Moq.Sequences
+{
+    public class SequenceException : Exception
+    {
+        public SequenceException(string message) : base(message)
+        {            
+        }
+    }
+}
