--- src/Properties/AssemblyInfo.cs
@@ -1,14 +1,0 @@
-﻿using System.Reflection;
-using System.Runtime.InteropServices;
-
-[assembly: AssemblyTitle("Moq.Sequences")]
-[assembly: AssemblyDescription("Supports checking that Moq setups are executed in sequence.")]
-[assembly: AssemblyProduct("Moq.Sequences")]
-[assembly: AssemblyCopyright("Declan Whelan © 2010")]
-[assembly: AssemblyCompany("Declan Whelan")]
-
-[assembly: ComVisible(false)]
-[assembly: Guid("4e1e90c1-9fdc-4aa1-be0a-46975884172e")]
-
-[assembly: AssemblyVersion("1.0.1.0")]
-[assembly: AssemblyFileVersion("1.0.1.0")]
