--- src/SequenceException.cs
@@ -1,11 +1,0 @@
-using System;
-
-namespace Moq.Sequences
-{
-    public class SequenceException : Exception
-    {
-        public SequenceException(string message) : base(message)
-        {            
-        }
-    }
-}
