--- src/Sequences.nuspec
@@ -1,19 +1,0 @@
-<?xml version="1.0"?>
-<package >
-  <metadata>
-    <id>$id$</id>
-    <version>$version$</version>
-    <authors>$author$</authors>
-    <owners>$author$</owners>
-    <licenseUrl>https://github.com/dwhelan/Moq-Sequences/blob/master/License.txt</licenseUrl>
-    <projectUrl>https://github.com/dwhelan/Moq-Sequences</projectUrl>
-    <requireLicenseAcceptance>false</requireLicenseAcceptance>
-    <description>$description$</description>
-    <releaseNotes>Added NuGet support.</releaseNotes>
-    <tags>Moq mock .net40</tags>
-    <dependencies>
-      <dependency id="Moq" version="4.0" />
-      <dependency id="NUnit" version="2.5" />
-    </dependencies>    
-  </metadata>
-</package>
