--- src/packages.config
@@ -1,5 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Moq" version="4.0.10827" />
-  <package id="NUnit" version="2.5.10.11092" />
-</packages>
